+++ mehari/src/db/dbsnp.rs
@@ -1,6 +1,6 @@
 use crate::common::Args as CommonArgs;
-use crate::db::PipelineConfig;
 use crate::db::keys::Var;
+use crate::db::{PipelineConfig, write_contig_dictionary};
 use crate::pbs::seqvars::DbsnpRecord;
 use anyhow::Error;
 use clap::Parser;
@@ -98,16 +98,7 @@ pub fn run(_common: &CommonArgs, args: &Args) -> Result<(), Error> {
     )?;
 
     tracing::info!("Writing dbSNP contig index metadata mapping into the meta CF...");
-    let options = rocksdb::Options::default();
-    let options = rocksdb_utils_lookup::tune_options(options, None);
-    let db = rocksdb::DB::open_cf(&options, &args.common.output, vec!["meta", "dbsnp"])?;
-    let cf_meta = db
-        .cf_handle("meta")
-        .ok_or_else(|| anyhow::anyhow!("meta CF not found"))?;
-
-    let map_guard = chrom_to_id.read().unwrap();
-    let serialized_dict = serde_json::to_vec(&*map_guard)?;
-    db.put_cf(&cf_meta, b"contig_dictionary", serialized_dict)?;
+    write_contig_dictionary(&args.common.output, "dbsnp", &chrom_to_id)?;
 
     Ok(())
 }
